Implemented Widget structure on Client webpage
