+++ mock banking page/index.html
@@ -51,6 +51,7 @@
                 </tr>
             </table>
         </div>
+        <div id="widget-container"></div>
     </div>
     <div id="right-container">
         <div class="note">
@@ -68,7 +69,6 @@
                 <li>Manage recordkeeping</li>
             </ul>
         </div>
-        <div id="widget-container"></div>
     </div>
 </div>
 <!-- Load React. -->
@@ -77,6 +77,6 @@
 <script src="https://unpkg.com/react-dom@17/umd/react-dom.development.js" crossorigin></script>
 
 <!-- Load our React component. -->
-<script src="../src/components/widget/widget.js"></script>
+<script src="widget/widget.js"></script>
 </body>
 </html>
