--- src/components/widget/widget.module.scss
@@ -1,20 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
-ul.react-multi-carousel-track li:nth-child(1) .ant-card:nth-child(1) {
-  margin-left: 20px;
-}
-ul.react-multi-carousel-track .ant-card {
-  margin: 0 5px;
-}
-
-button {
-  display: inline-block;
-  background-color: #0d647d;
-  color: #ffffff !important;
-  outline: none;
-  box-shadow: none;
-  border-radius: 6px;
-  min-width: 40px;
-}
